Refactor the codebase so that file names follow proper naming guidelines for Next.js. I also split the code for a lot of the files so that comoponents have their own files, types are defined into their own files, interfaces are defined into their own files, and I also adjusted the location of some of the files as well.
