+++ app/layout.tsx
@@ -2,12 +2,12 @@ import "./global.css";
 import type { Metadata } from "next";
 import { GeistSans } from "geist/font/sans";
 import { GeistMono } from "geist/font/mono";
-import { Navbar } from "./components/nav";
+import { Navbar } from "@/components/nav";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import Footer from "./components/footer";
+import Footer from "@/components/footer";
 import { baseUrl } from "./sitemap";
-import { PersonJsonLd } from "./components/seo/json-ld";
+import { PersonJsonLd } from "@/components/seo/json-ld";
 
 const siteTitle: string = "Mal Nushi";
 const siteDescription: string = "Mal Nushis Personal Website";
