+++ app/sidequests/legos/page.tsx
@@ -1,19 +1,9 @@
 import Link from "next/link";
 import type { Metadata } from "next";
 import React from "react";
-import LegoWordBuilder from "app/components/LegoWordBuilder";
-import TaggedBlogPosts from "app/components/TaggedBlogPosts";
-
-// Define the subpages for the legos page
-const legoSubpages = [
-  {
-    name: "Gallery",
-    slug: "legos_gallery",
-    description:
-      "I should charge you for this one, but I can't. Living in Charlotte is expensive man.",
-    externalUrl: "https://www.instagram.com/mal.builds.legos/",
-  },
-];
+import LegoWordBuilder from "@/components/lego-word-builder";
+import TaggedBlogPosts from "@/components/tagged-blog-posts";
+import { sidequestSubpages } from "@/data/sidequests/legos/subpages"; // Subpages defined here
 
 /**
  * Metadata for the Legos page. This object is used by Next.js to set the page title and description for SEO and browser display.
@@ -58,7 +48,7 @@ export default function LegosPage(): React.JSX.Element {
       {/*List of subpages for the lego section*/}
       <h2 className="mb-4 text-xl font-semibold tracking-tight">Sub-Topics</h2>
       <ul className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 mb-8">
-        {legoSubpages.map((subpage) => (
+        {sidequestSubpages.map((subpage) => (
           <li key={subpage.slug}>
             {/* This is a conditional render: use external link if exists, otherwise use internal link */}
             {subpage.externalUrl ? (
