+++ app/sidequests/pickleball/page.tsx
@@ -1,30 +1,8 @@
 import Link from "next/link";
 import type { Metadata } from "next";
-import PingPongText from "app/components/PingPongEffect";
-import TaggedBlogPosts from "app/components/TaggedBlogPosts";
-
-/**
- * Interface & class for the structure of pickleball subpages
- * @property {string} name - Display name of the subpage
- * @property {string} slug - URL slug for routing
- * @property {string} description - Short description of the subpage content
- */
-const pickleballSubpages: {
-  name: string;
-  slug: string;
-  description: string;
-}[] = [
-  {
-    name: "My Gear",
-    slug: "pickleball_gear",
-    description: "A look at my current setup.",
-  },
-  {
-    name: "Favorite Courts",
-    slug: "pickleball_courts",
-    description: "Best courts I have played in.",
-  },
-];
+import PingPongText from "@/components/pickleball-effect";
+import TaggedBlogPosts from "@/components/tagged-blog-posts";
+import { sidequestSubpages } from "@/data/sidequests/pickleball/subpages"; // Subpages defined here
 
 /**
  * Metadata for the Pickleball page. This object is used by Next.js to set the page title and description for SEO and browser display.
@@ -55,7 +33,7 @@ export default function PickleballPage(): React.JSX.Element {
 
       <h2 className="mb-4 text-xl font-semibold tracking-tight">Sub-Topics</h2>
       <ul className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 mb-8">
-        {pickleballSubpages.map((subpage) => (
+        {sidequestSubpages.map((subpage) => (
           <li key={subpage.slug}>
             <Link href={`/sidequests/pickleball/${subpage.slug}`}>
               <div className="bg-white dark:bg-slate-800 rounded-xl shadow p-4 hover:bg-blue-50 dark:hover:bg-gray-700 transition cursor-pointer h-full flex flex-col">
