+++ app/sidequests/skating/page.tsx
@@ -1,29 +1,7 @@
 import Link from "next/link";
-import SkateText from "app/components/SkateEffect";
-import TaggedBlogPosts from "app/components/TaggedBlogPosts";
-
-/**
- * Interface & class for the structure of skating subpages
- * @property {string} name - Display name of the subpage
- * @property {string} slug - URL slug for routing
- * @property {string} description - Short description of the subpage content
- */
-const skatingSubpages: {
-  name: string;
-  slug: string;
-  description: string;
-}[] = [
-  {
-    name: "My Gear",
-    slug: "skate_gear",
-    description: "A look at my current setup.",
-  },
-  {
-    name: "Favorite Spots",
-    slug: "skate_spots",
-    description: "Best places I've skated.",
-  },
-];
+import SkateText from "@/components/skate-effect";
+import TaggedBlogPosts from "@/components/tagged-blog-posts";
+import { sidequestSubpages } from "@/data/sidequests/skating/subpages"; // Subpages defined here
 
 /**
  * Metadata for the Skating page. This object is used by Next.js to set the page title and description for SEO and browser display.
@@ -54,7 +32,7 @@ export default function SkatingPage(): React.JSX.Element {
 
       <h2 className="mb-4 text-xl font-semibold tracking-tight">Sub-Topics</h2>
       <ul className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 mb-8">
-        {skatingSubpages.map((subpage) => (
+        {sidequestSubpages.map((subpage) => (
           <li key={subpage.slug}>
             <Link href={`/sidequests/skating/${subpage.slug}`}>
               <div className="bg-white dark:bg-slate-800 rounded-xl shadow p-4 hover:bg-blue-50 dark:hover:bg-gray-700 transition cursor-pointer h-full flex flex-col">
