+++ app/sidequests/skating/skate_gear/page.tsx
@@ -1,60 +1,6 @@
-import Link from "next/link";
 import type { Metadata } from "next";
-import React from "react";
-import { safeString } from "@/app/utils/safeString";
-
-/**
- * Interface representing a gear item
- * @property {string} name - Display name of the gear
- * @property {string} slug - URL slug for routing
- * @property {string} description - Short description of the gear
- * @property {string} externalUrl - Optional external URL if the item links to an external site
- */
-interface GearItem {
-  name: string;
-  slug: string;
-  description: string;
-  externalUrl?: string;
-}
-
-/**
- * Array of gear items that represent subpages (or external links)
- * Each item includes a name, slug, description, and optionally an external URL.
- */
-const GearSubpages: GearItem[] = [
-  {
-    name: "The Original GRLSWIRL x Carver Surfskate",
-    slug: "grlswirl_x_carver",
-    description: "My very first surfskate and my daily rider.",
-  },
-  {
-    name: 'Penny Ocean Mist 22" Penny Board',
-    slug: "pennyboard_ocean_mist",
-    description:
-      "Bought this bad boy for traveling. Never once took it for traveling.",
-  },
-  {
-    name: "Powell / Peralta / Rodriguez Skull and Sword Skateboard",
-    slug: "powell_peralta",
-    description:
-      "A cool 80's style board that is configured as a skatepark board, but also a cruiser.",
-  },
-  {
-    name: "Fila Legacy Pro 84 Roller Blade",
-    slug: "fila_legacy_pro",
-    description: "My very first pair of roller skates (thank you Sally).",
-  },
-  {
-    name: "187 Killer Pads 6-Pack Set",
-    slug: "killer_pads",
-    description: "Can never be too careful.",
-  },
-  {
-    name: "Protec Classic Skateboard Helmet CPSC",
-    slug: "protec_helmet",
-    description: "Can most definitely never be too careful.",
-  },
-];
+import { GearSubpages } from "@/data/sidequests/skating/gear-subpages"; // Subpages for each gear item
+import { GearListPage } from "@/components/gear/gear-list-page";
 
 // SEO metadata
 export const metadata: Metadata = {
@@ -62,96 +8,14 @@ export const metadata: Metadata = {
   description: "Details about my skateboard setup and accessories.",
 };
 
-/**
- * GearCard component
- *
- * A presentational component that displays the visual card for a single gear item.
- * It renders a styled container with the gear's name and description.
- *
- * This component does not handle linking — it is meant to be wrapped by either
- * a Next.js `<Link>` for internal navigation or an `<a>` tag for external links.
- *
- * @param {object} props - Component props
- * @param {string} props.name - The display name of the gear item
- * @param {string} props.description - A short description of the gear item
- *
- * @returns {React.JSX.Element} A styled card element showing the gear details
- *
- * @example
- * <GearCard
- *   name="Penny Ocean Mist 22\" Penny Board"
- *   description="Bought this bad boy for traveling. Never once took it for traveling."
- * />
- */
-function GearCard({
-  name,
-  description,
-}: {
-  name: string;
-  description: string;
-}) {
+// Display the gear items
+export default function GearPage() {
   return (
-    <div className="bg-white dark:bg-slate-800 rounded-xl shadow p-4 hover:bg-blue-50 dark:hover:bg-gray-700 transition cursor-pointer h-full flex flex-col">
-      <h3 className="text-md font-semibold mb-1">{safeString(name)}</h3>
-      <p className="text-xs text-neutral-500 dark:text-neutral-300 flex-grow">
-        {safeString(description)}
-      </p>
-    </div>
-  );
-}
-
-/**
- * GearPage component.
- *
- * The main page for displaying a grid of gear items.
- * It pulls gear data from `GearSubpages` and renders a list of `GearCard` components.
- * Each card is wrapped in either a Next.js `<Link>` for internal navigation or an `<a>` tag for external links.
- *
- * @returns {React.JSX.Element} The rendered gear page containing the gear list.
- */
-export default function GearPage(): React.JSX.Element {
-  const pageTitle = "My Skating Gear";
-  const pageDescription = "Details about my skateboard setup and accessories.";
-
-  return (
-    <section>
-      <h1 className="mb-2 text-2xl font-semibold tracking-tighter">
-        {pageTitle}
-      </h1>
-      <p className="mb-8 text-neutral-600 dark:text-neutral-400">
-        {pageDescription}
-      </p>
-
-      <ul className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 mb-8">
-        {GearSubpages.map((subpage) => {
-          const external = safeString(subpage.externalUrl);
-          const href =
-            external ||
-            `/sidequests/skating/skate_gear/${encodeURIComponent(
-              subpage.slug
-            )}`;
-
-          return (
-            <li key={subpage.slug}>
-              {external ? (
-                <a href={href} target="_blank" rel="noopener noreferrer">
-                  <GearCard
-                    name={subpage.name}
-                    description={subpage.description}
-                  />
-                </a>
-              ) : (
-                <Link href={href}>
-                  <GearCard
-                    name={subpage.name}
-                    description={subpage.description}
-                  />
-                </Link>
-              )}
-            </li>
-          );
-        })}
-      </ul>
-    </section>
+    <GearListPage
+      title="My Skating Gear"
+      description="Details about my skateboard setup and accessories."
+      basePath="/sidequests/skating/skate_gear"
+      items={GearSubpages}
+    />
   );
 }
