+++ components/flashing-text.tsx
