+++ components/typing-effect.tsx
