+++ data/sidequests/legos/subpages.ts
@@ -1,0 +1,11 @@
+import type { SidequestSubpage } from "@/interfaces/sidequest-subpages";
+
+export const sidequestSubpages: SidequestSubpage[] = [
+  {
+    name: "Gallery",
+    slug: "legos_gallery",
+    description:
+      "I should charge you for this one, but I can't. Living in Charlotte is expensive man.",
+    externalUrl: "https://www.instagram.com/mal.builds.legos/",
+  },
+];
