+++ data/sidequests/music/gear-subpages.ts
@@ -1,0 +1,36 @@
+import type { GearItem } from "@/interfaces/gear-item";
+
+export const GearSubpages: GearItem[] = [
+  {
+    name: "Yamaha F335 Acoustic Guitar",
+    slug: "y_f335",
+    description: "Far from the best acoustic guitar, but hey, it works for me.",
+  },
+  {
+    name: "Epiphone ES-335 Semi-Hollowbody Electric Guitar",
+    slug: "es-335",
+    description:
+      "The very first guitar pops ever gave to me. I cherish it like it is my kid.",
+  },
+  {
+    name: "PRS SE Custom 24 Solid-Body Electric Guitar",
+    slug: "prs_se_custom_24",
+    description: "Very versatile and so comfortable to play.",
+  },
+  {
+    name: "Yamaha RBX170 Electric Bass",
+    slug: "y_rbx170",
+    description:
+      "The very first bass pops ever gave to me. Cheaper than my Viola bass, but packs a mean punch.",
+  },
+  {
+    name: "Epiphone Viola Bass",
+    slug: "e_viola",
+    description: "Honestly, I am just a Paul McCartney fan.",
+  },
+  {
+    name: "Epiphone MM-20 Acoustic Mandolin",
+    slug: "e_mm-20",
+    description: "Talk about blisters on your fingers...",
+  },
+];
