+++ data/sidequests/music/subpages.ts
@@ -1,0 +1,11 @@
+import type { SidequestSubpage } from "@/interfaces/sidequest-subpages";
+
+export const sidequestSubpages: SidequestSubpage[] = [
+  {
+    name: "My Gear",
+    slug: "music_gear",
+    description:
+      "The gear responsible for the sounds that make my neighbors think a rat’s meeting its end.",
+    externalUrl: "",
+  },
+];
