+++ data/sidequests/photography/subpages.ts
@@ -1,0 +1,15 @@
+import type { SidequestSubpage } from "@/interfaces/sidequest-subpages";
+
+export const sidequestSubpages: SidequestSubpage[] = [
+  {
+    name: "My Gear",
+    slug: "photography_gear",
+    description: "The gear that makes me look like I know what I am doing.",
+  },
+  {
+    name: "Gallery",
+    slug: "photography_gallery",
+    description: "$0 entry to the 'Mal Nushi Museum of Photography'.",
+    externalUrl: "https://www.instagram.com/malspixelempire/",
+  },
+];
