+++ data/sidequests/pickleball/gear-subpages.ts
@@ -1,0 +1,13 @@
+import type { GearItem } from "@/interfaces/gear-item";
+
+/**
+ * Array of gear items that represent subpages (or external links)
+ * Each item includes a name, slug, description, and optionally an external URL.
+ */
+export const GearSubpages: GearItem[] = [
+  {
+    name: "Vatic Pro Prism Flash Carbon Fiber 16MM - Foam Injected Walls",
+    slug: "vatic_pro_prism",
+    description: "I slice well with this bad boy.",
+  },
+];
