+++ data/sidequests/pickleball/subpages.ts
@@ -1,0 +1,14 @@
+import type { SidequestSubpage } from "@/interfaces/sidequest-subpages";
+
+export const sidequestSubpages: SidequestSubpage[] = [
+  {
+    name: "My Gear",
+    slug: "pickleball_gear",
+    description: "A look at my current setup.",
+  },
+  {
+    name: "Favorite Courts",
+    slug: "pickleball_courts",
+    description: "Best courts I have played in.",
+  },
+];
