+++ data/sidequests/sidequests.ts
@@ -1,0 +1,12 @@
+import type { Sidequest } from "@/interfaces/sidequest";
+
+export const sidequests: Sidequest[] = [
+  { name: "Skating", emoji: "🛹", slug: "skating" },
+  { name: "Pickleball", emoji: "🏓", slug: "pickleball" },
+  { name: "Photography", emoji: "📷", slug: "photography" },
+  { name: "Legos", emoji: "🧱", slug: "legos" },
+  { name: "Music", emoji: "🎵", slug: "music" },
+  { name: "Keyboards", emoji: "⌨️", slug: "keyboards" },
+  { name: "Travel", emoji: "✈️", slug: "travel" },
+  { name: "Design", emoji: "🎨", slug: "design" },
+];
