+++ data/sidequests/skating/subpages.ts
@@ -1,0 +1,14 @@
+import type { SidequestSubpage } from "@/interfaces/sidequest-subpages";
+
+export const sidequestSubpages: SidequestSubpage[] = [
+  {
+    name: "My Gear",
+    slug: "skate_gear",
+    description: "A look at my current setup.",
+  },
+  {
+    name: "Favorite Spots",
+    slug: "skate_spots",
+    description: "Best places I've skated.",
+  },
+];
