+++ interfaces/sidequest.ts
@@ -1,0 +1,5 @@
+export interface Sidequest {
+  name: string;
+  emoji: string;
+  slug: string;
+}
