+++ types/supported-languages.ts
@@ -1,0 +1,29 @@
+/**
+ * A union type representing all supported programming languages.
+ */
+export type SupportedLanguage =
+  | "AppleScript"
+  | "TypeScript"
+  | "JavaScript"
+  | "Python"
+  | "HTML"
+  | "CSS"
+  | "Java"
+  | "C"
+  | "C++"
+  | "C#"
+  | "Ruby"
+  | "Go"
+  | "Rust"
+  | "PHP"
+  | "Swift"
+  | "Kotlin"
+  | "Dart"
+  | "Bash"
+  | "Shell"
+  | "JSON"
+  | "YAML"
+  | "Markdown"
+  | "SQL"
+  | "R"
+  | "Lua";
