+++ utils/safe-string.ts
